# api/goa/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: DVC Points Calculator
    description: Backend service for DVC Points Calculator UI
    version: "1.0"
servers:
    - url: http://localhost:8080
paths:
    /chart/{resortCode}/{year}:
        get:
            tags:
                - Points
            summary: GetPointChart Points
            operationId: Points#GetPointChart
            parameters:
                - name: resortCode
                  in: path
                  description: the resort's code
                  required: true
                  schema:
                    type: string
                    description: the resort's code
                    example: ssr
                    pattern: '[a-z]{3}'
                  example: ssr
                - name: year
                  in: path
                  description: the year
                  required: true
                  schema:
                    type: integer
                    description: the year
                    example: 2019
                    minimum: 1990
                    maximum: 2100
                  example: 2094
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PointChart'
                            example:
                                code: ssr
                                resort: Disney's Saratoga Springs Resort & Spa
                                roomTypes:
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                tiers:
                                    - dateRanges:
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                      roomTypePoints:
                                        Ab quaerat delectus amet.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                                        Blanditiis eos.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                                        Dicta ipsum.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                                    - dateRanges:
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                        - endDate: 1-15
                                          startDate: 1-15
                                      roomTypePoints:
                                        Ab quaerat delectus amet.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                                        Blanditiis eos.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                                        Dicta ipsum.:
                                            weekday: 6454418589515295321
                                            weekend: 3368857188488503470
                "404":
                    description: 'not_found: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /resort/{resortCode}:
        get:
            tags:
                - Points
            summary: GetResort Points
            operationId: Points#GetResort
            parameters:
                - name: resortCode
                  in: path
                  description: the resort's code
                  required: true
                  schema:
                    type: string
                    description: the resort's code
                    example: ssr
                    pattern: '[a-z]{3}'
                  example: ssr
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                description: 'Response body may alternatively be one of #/components/schemas/ResortResultResponseResortOnly or #/components/schemas/ResortResultResponseResortUpdate'
                                $ref: '#/components/schemas/ResortResult'
                            example:
                                code: ssr
                                name: Disney's Saratoga Springs Resort & Spa
                                roomTypes:
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                "404":
                    description: 'not_found: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /resort/{resortCode}/year/{year}:
        get:
            tags:
                - Points
            summary: GetResortYear Points
            operationId: Points#GetResortYear
            parameters:
                - name: resortCode
                  in: path
                  description: the resort's code
                  required: true
                  schema:
                    type: string
                    description: the resort's code
                    example: ssr
                    pattern: '[a-z]{3}'
                  example: ssr
                - name: year
                  in: path
                  description: the year
                  required: true
                  schema:
                    type: integer
                    description: the year
                    example: 2034
                    minimum: 1990
                    maximum: 2100
                  example: 2011
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResortYearResult'
                            example:
                                code: ssr
                                name: Disney's Saratoga Springs Resort & Spa
                                roomTypes:
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                    - bedrooms: 1
                                      beds: 2
                                      code: 1bp
                                      name: One Bedroom Villa (Preferred View)
                                      sleeps: 5
                                year: 2068
                "404":
                    description: 'not_found: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
    schemas:
...
        PointChart:
            type: object
            properties:
                code:
                    type: string
                    description: resort's code
                    example: ssr
                    pattern: '[a-z]{3}'
                resort:
                    type: string
                    description: resort's code
                    example: Disney's Saratoga Springs Resort & Spa
                roomTypes:
                    type: array
                    items:
                        $ref: '#/components/schemas/RoomType'
                    example:
                        - bedrooms: 1
                          beds: 2
                          code: 1bp
                          name: One Bedroom Villa (Preferred View)
                          sleeps: 5
                        - bedrooms: 1
                          beds: 2
                          code: 1bp
                          name: One Bedroom Villa (Preferred View)
                          sleeps: 5
                        - bedrooms: 1
                          beds: 2
                          code: 1bp
                          name: One Bedroom Villa (Preferred View)
                          sleeps: 5
                        - bedrooms: 1
                          beds: 2
                          code: 1bp
                          name: One Bedroom Villa (Preferred View)
                          sleeps: 5
                tiers:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tier'
                    example:
                        - dateRanges:
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                          roomTypePoints:
                            Ab quaerat delectus amet.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
                            Blanditiis eos.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
                            Dicta ipsum.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
                        - dateRanges:
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                            - endDate: 1-15
                              startDate: 1-15
                          roomTypePoints:
                            Ab quaerat delectus amet.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
                            Blanditiis eos.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
                            Dicta ipsum.:
                                weekday: 6454418589515295321
                                weekend: 3368857188488503470
            example:
                code: ssr
                resort: Disney's Saratoga Springs Resort & Spa
                roomTypes:
                    - bedrooms: 1
                      beds: 2
                      code: 1bp
                      name: One Bedroom Villa (Preferred View)
                      sleeps: 5
                    - bedrooms: 1
                      beds: 2
                      code: 1bp
                      name: One Bedroom Villa (Preferred View)
                      sleeps: 5
                tiers:
                    - dateRanges:
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                      roomTypePoints:
                        Ab quaerat delectus amet.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
                        Blanditiis eos.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
                        Dicta ipsum.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
                    - dateRanges:
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                        - endDate: 1-15
                          startDate: 1-15
                      roomTypePoints:
                        Ab quaerat delectus amet.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
                        Blanditiis eos.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
                        Dicta ipsum.:
                            weekday: 6454418589515295321
                            weekend: 3368857188488503470
            required:
                - code
                - resort
                - roomTypes
                - tiers
...
        ResortResultResponseResortOnly:
            type: object
            properties:
                code:
                    type: string
                    description: resort's code
                    example: ssr
                    pattern: '[a-z]{3}'
                name:
                    type: string
                    description: resort's name
                    example: Disney's Saratoga Springs Resort & Spa
            description: ResortResultResponse result type (resortOnly view)
            example:
                code: ssr
                name: Disney's Saratoga Springs Resort & Spa
...
        ResortResultResponseResortUpdate:
            type: object
            properties:
                name:
                    type: string
                    description: resort's name
                    example: Disney's Saratoga Springs Resort & Spa
            description: ResortResultResponse result type (resortUpdate view)
            example:
                name: Disney's Saratoga Springs Resort & Spa
...